Take condition expression in EncFilter to fix Catalyst bug
When EncFilter only took an opcode, which is opaque to Spark SQL, the
ColumnPruning optimizer rule would assume it did not need any input
columns and would therefore drop columns that it really did need as long
as they were not needed by later plan nodes. To avoid this, we have to
accept a filter expression that references all needed columns.