Store rows in blocks to reduce splitting/merging
Splitting the output of an enclave call into rows and merging it back
together for the next enclave call turned out to be very expensive. This
commit makes the Opaque operators default to passing rows around as
blocks, only merging or splitting when necessary for input/output or for
non-Opaque operators.