Use row-specific upper bound when writing blocks
This makes each row more compact when written out so there are fewer
bytes to encrypt. However, it actually hurts performance for hardware
mode, because we pack many rows into each sort buffer, but the rows then
expand to use ROW_UPPER_BOUND enclave memory each. This causes the sort
working set to exceed L3 cache, forcing a lot of writes to main memory
which have to be encrypted.