std::vector -> placement new to avoid invoking copy constructor
The default copy constructors of NewRecord and NewJoinRecord are wrong,
because they copy over the row pointer but then free it as a result of
calling the destructor on the old copy.

These copy constructors were being invoked by the std::vector
constructor. causing a segfault when the row pointers were accessed.

We therefore use placement new to initialize the record data rather than
std::vector. It would have been ideal to use std::vector::emplace, but
that is new in C++11 and the SGX SDK only provides the C++03 STL. We
could also have defined a copy constructor for the record types that
sets the row pointer from the old record to NULL to prevent it from
being freed.