Expression-generic Spark SQL operators for Opaque
Instead of having to pass an opcode for each operator, the user supplies
the Catalyst expression and we pattern match on it to find the matching
opcode. In addition to being a better interface, this makes it safer to
write queries, since tables are checked for conformance with the
expected schema before being passed to the enclave.