Pass last row to next partition in aggregate
This is necessary because the last aggregated group does not necessarily contain
enough information to tell whether a given row is part of the same group. This
occurs when spark.sql.retainGroupColumns is false.

To determine whether to continue aggregating into the last partial aggregate
from the previous partition or to begin with an empty aggregate, we need the
last row of the previous partition.