Store temporary rows safely
The row pointer returned by EncryptedBlocksToRowReader#next() is temporary and
may be invalidated when next() is called again. However, current operator
implementations store the returned pointers and unsafely dereference them later.
This leads to crashes when processing input with multiple encrypted blocks.

This commit adds FlatbuffersTemporaryRow, a convenient way to store such rows
safely. Row pointers are copied into temporary storage and a pointer to the
copied row is returned.