Fix incorrect attribute references in EncryptedJoin
When generating a plan for EncryptedJoin, OpaqueOperators emits incorrect
SortOrders for some schemas. The SortOrders reference the attributes
corresponding to the input keys rather than to their aliases generated by
tagForJoin. This causes the enclave to compare the wrong attributes when sorting
for joins.

For example, when joining [a, b, c] with [d, e] on a == e, the join should
proceed as follows:

1. tagForJoin: Project [a, b, c] to [Literal(0) AS _tag, a AS key_0, a, b, c].
2. tagForJoin: Project [d, e] to [Literal(1) AS _tag, e AS key_0, d, e].
3. Union these two tables.
4. Sort on (_tag, key_0).
5. Sort-merge join.
6. Drop _tag and key_0 to produce the output.

Instead, step 4 incorrectly sorts on (_tag, a). Since the enclave implements
attribute references positionally, this actually results in sorting on (_tag, a
|| d), which will generate incorrect results and may result in an enclave crash
due to mismatched field types.

This commit fixes the attribute references by ensuring that the SortOrders
reference the aliases generated by tagForJoin.