Handle enclave errors using C++ exceptions rethrown to Java
Enclave errors are currently signaled using `ocall_exit()` (often invoked using
the `check()` macro). This quits the entire JVM process including the Spark
worker or driver, which is not user-friendly.

JNI provides a way to throw Java exceptions, but the corresponding JNI
call (`JNIEnv::ThrowNew()`) only throws the Java exception once the JNI method
invocation returns. That is, `ThrowNew()` returns to the caller, which is
presumably in a corrupt state and cannot continue normal execution. This
prevents `ThrowNew()` from being used as a direct replacement for
`ocall_exit()`.

Instead, this commit uses C++ exceptions within the enclave to transfer control
flow to a handler at the top level of each ecall. The handler invokes
`ThrowNew()` (via `ocall_throw()`, a newly-introduced function) and directly
returns from the ecall.

Since it is no longer desirable to use `ocall_exit()`, this commit also removes
the `check()` macro and replaces all uses of it with C++ exception throws.