Fix enclave startup crash in SBT Release mode with SGX_MODE=HW
The `debug` argument to `sgx_create_enclave` should always be 1 unless a special
process with Intel has been completed [1].

However, CMake's Release mode passes -DNDEBUG, which causes it to be 0, as
defined in sgx_urts.h [2]. This causes a crash at enclave startup when
SGX_MODE=HW is passed.

This commit fixes this by also passing -DEDEBUG in CMake's Release mode to match
the behavior of the SGX SDK's "pre-release" mode. This causes the `debug`
argument to be 1 in this mode, fixing the crash.

Fixes #75.

[1] "Enclaves compiled under the Release profile will not work until the
developer completes the production licensing process." [source]

[2] https://github.com/intel/linux-sgx/blob/6a280c11741244160e7dec4f05211f5ae5438b8d/common/inc/sgx_urts.h#L82