Fix unsafe use of Flatbuffers temporary pointers (#83)
In ExpressionEvaluator, temporary pointers may be invalidated not only by
directly writing to builder, which can cause a reallocation, but indirectly
through a recursive call to eval_helper. The current code fails to account for
this, resulting in rare data corruption issues. A similar problem is present in
the other evaluators.

This PR solves the problem by preferring to store offsets, which cannot be
invalidated except by clearing the builder, and delaying the conversion from
offsets to pointers until all recursive expression evaluation is complete.