Implement join conditions
Spark SQL sometimes tries to combine a filter operator and a join operator by
putting the filter predicate into the join condition using the
PushPredicateThroughJoin rule. Opaque didn't fully implement join conditions, so
in this case the filter would get ignored. This commit implements join
conditions using an EncryptedFilter on the result of the join.