Complete implementation of remote attestation (#88)
This PR completes the implementation of remote attestation, adding code to communicate with the Intel Attestation Service and to securely transfer the shared data encryption key (`edu.berkeley.cs.rise.opaque.Utils.sharedKey`) to the enclaves. The Intel Attestation Service request code is adapted from [Intel's sample](https://github.com/intel/sgx-ra-sample).

Attestation is mediated by the remote attestation service provider (`src/enclave/ServiceProvider/ServiceProvider.h`), which communicates with both the Intel Attestation Service and the client (the Spark driver) to verify the worker enclaves. This PR refactors the service provider to expose an object-oriented interface.

Because remote attestation is now used to transfer the shared key to the enclaves, it is required for running any Opaque queries. It is therefore invoked whenever `Utils.initSQLContext` is called, including in tests. However, an attestation failure is ignored unless the `OPAQUE_REQUIRE_ATTESTATION` environment variable is set. This allows running Opaque in simulation mode.