Robust attestation (#170)
This PR contains a more fault tolerant way to do attestation. The remote attestation code runs in a loop in a separate thread in order to constantly check for unattested machines. This is done through the use of an accumulator variable -- every time a new enclave is started, the accumulator variable's state is changed to indicate that attestation needs to be re-run.

One caveat is that this attestation design relies on Spark's fault tolerance design. If a task attempts to run on an executor with an unattested enclave, the code will throw an exception. Therefore, the `spark.task.maxFailures` must be set to a number greater than 1 so that Spark will rerun the task. Currently this configuration variable is set to 10 for the tests.