Fix Flatbuffer serialization exceeding 2gb and improve performance logging (#172)
This PR fixes #158.

The solution establishes a method to retry Flatbuffer serialization if an assertion error occurs. In this case, MaxBlockSize is decreased by an order of magnitude before the serialization is attempted again.

This PR also includes improvements to opaque operator logging. Previously, most operators were by default caching intermediate DataFrames in order to accurately measure the amount of time it took for the operators to run. This was resulting in OOM issues, since in practice these intermediate DFs could be much larger than the final result. The fix involves specifying an additional parameter to the benchmark code. Note that for larger datasets, it is possible that benchmarking would fail when this parameter is set for the reason mentioned.