Query Listener (#8)
This PR creates a Listener object with a main class that is responsible for creating a gRPC server that expects to receive queries in the form of strings. With these query strings, another object OpaqueILoopHandler handles interfacing with OpaqueILoop, a thin wrapper over SparkILoop. With OpaqueILoopHandler, arbitrary strings that contain Scala code can easily be evaluated with a programmatic instance of a Scala REPL. Variables are persistent across multiple calls to OpaqueILoopHandler.run, which makes extending this to an interactive case much easier.