Listener support for submitting through Spark (#12) (#195)
This PR includes the necessary changes to have the Listener class run correctly when submitted to a formal Spark cluster.

We need to provide the fat jar created by build/sbt test:assembly with a newer version of the google.guava library. This is because the gRPC library did not work with the one provided by Spark, which was 14.0. This involves adding a shading rule that changes all previous classpaths prepending by com.google to the new library included in libraryDependencies.
The Spark shell had its own custom formatting that was causing results received on the client side to include some parsing rules that were not being evaluated. There is now a line in the docs mentioning to set "-Dscala.color" on starting up the JVM, which is sbt's default.
We had to explicitly add the jars provided to Spark to the new IMain's classpath, otherwise the Spark library could not be found.
Also includes updates to docs to reflect these changes.