Better assembly build (#198)
This PR enables running the listener with build/sbt assembly instead of build/sbt test:assembly (basically submitting a jar to Spark that does not contain unnecessary test files). It also renames the shade rule to better reflect what we are doing (changing references to Guava classes for non-Spark dependencies to use a newer version). We actually do not have to selectively pick the libraries this should be done for, since Spark is marked as "Provided" automatically by the sbt-spark plugin, meaning it is not included in the fat jar created by build/sbt assembly.