Revamp Documents (#208)
Resolves https://github.com/mc2-project/opaque/issues/205

This splits up the usage page in the docs into two: running with a trusted driver, and running with an untrusted one (using MC2 Client). It also creates a default key in `initOpaqueSQL` (previous `initSQLContext`), that would need to be set by loading in a key file from the config later.