Add Pyspark listener (#248)
This PR adds support for a Pyspark listener that is intended to receive queries from the MC2 Client. It contains SBT code to generate Python protobuf files and package them into a zip automatically for submitting to a standalone Spark cluster. The instructions for running the listener are in the client docs.

Unfortunately, it also introduces a Python dependency to Opaque SQL, since there are required packages to start the Python listener service.